Views support for Shindig.
This implements SHINDIG-41, conforming to the canonical Gadget XSD. View ID semantics are
that view IDs are split on commas, and each view ID found appends to the content for that view.
This allows for code sharing between views and provides a relatively clean implementation.

A blank or nonexistent view maps to ID "default".

Changes are made to be backwards-compatible, maintaining the old parameter-less
GadgetSpec.getContentData() API. TBD: whether to deprecate this method.

Also: Updating pom.*xml to Caja r422 (further updates should be done soon).



git-svn-id: https://svn.apache.org/repos/asf/incubator/shindig/trunk@617309 13f79535-47bb-0310-9956-ffa450edef68