The GadgetDataServlet now fetches the security token from all requests and creates a GadgetToken using the gadgetsigner (this is the same way that the signed request methods make gadget tokens)
This token is now passed to the people service methods (which have todos to refactor the apis) in order to determine the current viewer and owner.

For the sample container, this now means that you can change the current owner and viewer on the page. It also means that the state file format has changed. You no longer declare the owner and viewer, rather you declare people with friends. The dtd has been updated to reflect some of this. 



git-svn-id: https://svn.apache.org/repos/asf/incubator/shindig/trunk@638380 13f79535-47bb-0310-9956-ffa450edef68