Initial version of REST API support for PHP Shindig.
This isn't complete yet, so don't try to put this into production just yet ok :)

To try this out, edit your container.js set useRestful : true and set the social data 
path to /social/rest.

To make sure we don't break anyone's live code during the development of the REST interface
i've seperated out the old wireformat and rest format into 2 different directories: 
src/socialdata = the un-touched original wireformat code
src/socialrest = the new rest code

Unfortunately the internal API's had to be changed to make everything fit with the REST API,
the bigest change is that i moved away from the single handler class and went (back) to the
3 Data Service classes (People, AppData and Activities). 

The format is close enough to the old one though that it shouldn't take a lot of effort to
port your code to the new interface once your ready to switch over.



git-svn-id: https://svn.apache.org/repos/asf/incubator/shindig/trunk@668005 13f79535-47bb-0310-9956-ffa450edef68