Added url templates to the dataservice handlers to simplify their code a bunch. This logic went into the RequestItem class.
This also handles the spec format for posts now. For creating activities and app data the postdata is read directly instread of being mapped under an "entry" item. This changed the js code so I wrapped the changes in a useLegacy flag. Once the php updates to the more spec exact format we will remove this flag. 




git-svn-id: https://svn.apache.org/repos/asf/incubator/shindig/trunk@670107 13f79535-47bb-0310-9956-ffa450edef68