New http caching header functionality for the proxy service. This is still somewhat experimental, and while it has been tried in all the major browsers it hasn't been tested with every gadget yet.
What it does is add propper http header handling for e-tags and if-modified-since
headers, where it:

- always ignores the remote party's etag and generate our own (md5(content))
- use the remote party's last-modified if it's available, otherwise
  use the time that the request was last made.
- If the e-tag matches, 304 is returned and if-modified-since is not evaluated
  else it checks the if-modified-since

The idea is that most modern browsers support e-tags just fine, so that's our
prefered mechanism for returning 304 Not Modified to requests. However there
are proxy servers (and old browsers) that do not support e-tags, however they
do support last-modified / if-modified-since, so that's our fallback mechanism.

To further try to improve the cachability of the proxied requests, the code
also sets a Cache-Control:public, maxAge={refreshInterval} header, and a
Expires : (time + refreshInterval) header, as to save the round trip on things
that should be cachable.

The refreshInterval used is either the one specified in the request (&refresh=<seconds>)
and if non is set, the global caching time policy is used. (One day in the default
config).

Any and all feedback is very welcome!



git-svn-id: https://svn.apache.org/repos/asf/incubator/shindig/trunk@673461 13f79535-47bb-0310-9956-ffa450edef68