SHINDIG-427 - Add Content-Rewrite support. Big thanks to Jorge Condomi & Gonzalo Aune for their hard work on this.
This patch impements support for the optional content-rewrite feature,
including support for include-urls, exclude-urls and include and exclude
tags. 

Usage (to exclude everything):
	<Optional feature="content-rewrite">
 		<Param name="exclude-urls">.*</Param>
	</Optional>

Usage to include everything:
	<Optional feature="content-rewrite">
  		<Param name="include-urls">.*</Param>
	</Optional>

See javascript/samplecontainer/examples/rewriter{on,off}.xml as an
example on how to use this.

Rewriting by default has been disabled since this was causing trouble
with quite a few gadgets, so this is an opt-in only feature, however
gadget developers are strongly encouraged to use this since it helps
reduce the load on their servers & latency a lot.

Note: this doesn't currently rewrite the content of included css/js files
since there's no gadget feature state available during the proxy request,
this is something we might change in the future.

This patch also sneaks in a few fixes to the FilesServlet and HttpServlet
classes to optimize the browser caching behaviour.



git-svn-id: https://svn.apache.org/repos/asf/incubator/shindig/trunk@675660 13f79535-47bb-0310-9956-ffa450edef68