This patch adds logic to decode incomming data structures according to the Content-Type header.
It expects a "Content-Type: application/atom+xml" for Atom input,
and a "Content-Type: application/json" header for json input.

If no Content-Type header is provided, it will assume it's json.

The decoding logic has been moved out of the RestServlet and into
the converters/Input*Converter classes and the logic is now consitent
with the output converters.

The actual atom -> internal structure hoisting still remains to be
done & will follow soon.

ps this also slipstreams in a fix to JsLibrary, since the caja feature
was being picked up by it and the res:// reference caused it to error
out, to fix this for the time being it now ignores any res:// references
and represents it as 'empty content'



git-svn-id: https://svn.apache.org/repos/asf/incubator/shindig/trunk@678138 13f79535-47bb-0310-9956-ffa450edef68