SHINDIG-679
Patch from Ben Smith
The change in org.apache.shindig.social.opensocial.model.BodyType, turning the weight and height attributes to floats, was not reflected in org.apache.shindig.social.opensocial.jpa.BodyTypeDB (in the samples project).

Requires changing the attribute and getter/setter methods to use float, and the PopulatePerson test helper to send floats instead of strings.

Thanks

(Sorry should have caught this one... as it was me who made the change in the first place)


git-svn-id: https://svn.apache.org/repos/asf/incubator/shindig/trunk@711833 13f79535-47bb-0310-9956-ffa450edef68