Compatibility with Yahoo and NetFlix OAuth APIs.
For Yahoo: provide a mechanism to disable sending opensocial params for 
pure OAuth requests, since their request and access token endpoints are
strict on that topic.

For both Yahoo and Netflix: provide a mechanism to request extra data
returned from the access token URL.  Both Yahoo and Netflix return a 
user-id along with the access token, and require that you specify the 
user-id in subsequent API calls.



git-svn-id: https://svn.apache.org/repos/asf/incubator/shindig/trunk@712613 13f79535-47bb-0310-9956-ffa450edef68