Rolling back rpc.js to r764984.
In the past week and a half, we have had significant pain integrating the updated (RMR, early-message queueing, refactored) rpc.js into our existing deployments. The process has been extremely illuminating, highlighting the need for greatly-improved end-to-end testing of every conceivable configuration and use of this library.

I am temporarily rolling back the change to take an opportunity to review the improved version in its entirety and message some of the integration challenges we've faced - and that others may as well (the one that was ultimately most difficult for us: some containers chose to cache a many-versions-old rpc.js that, for subtle reasons, proved incompatible w/ the latest).



git-svn-id: https://svn.apache.org/repos/asf/incubator/shindig/trunk@782853 13f79535-47bb-0310-9956-ffa450edef68